Add files via upload
This fast generator is a simple interesting project which is simply in HTML, CSS, and JavaScript. Here, the user has to click on Generate password button and the password will appear with different characters included. You can copy the generated password to the clipboard. Also, this project includes a lot of JavaScript for making validations to certain parts of the project.